# .synnovator/openapi.yaml
openapi: 3.0.3
info:
  title: Synnovator API
  version: 1.0.0
  description: API for Synnovator - Activity and Competition Management Platform
  contact:
    name: API Support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
- url: http://localhost:8000
  description: Development server
- url: https://api.synnovator.com
  description: Production server
...
paths:
  /auth/login:
    post:
      summary: User login
      operationId: login
      tags:
      - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      summary: User logout
      operationId: logout
      tags:
      - auth
      security:
      - bearerAuth: []
      responses:
        '204':
          description: Logged out successfully
  /auth/refresh:
    post:
      summary: Refresh access token
      operationId: refresh_token
      tags:
      - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /groups:
    get:
      summary: List groups
      operationId: list_groups
      tags:
      - groups
      parameters:
      - $ref: '#/components/parameters/SkipParam'
      - $ref: '#/components/parameters/LimitParam'
      - name: visibility
        in: query
        schema:
          $ref: '#/components/schemas/GroupVisibility'
      responses:
        '200':
          description: List of groups
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedGroupList'
    post:
      summary: Create group
      operationId: create_group
      tags:
      - groups
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupCreate'
      responses:
        '201':
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '422':
          $ref: '#/components/responses/ValidationError'
  /groups/{group_id}:
    get:
      summary: Get group
      operationId: get_group
      tags:
      - groups
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Group details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update group
      operationId: update_group
      tags:
      - groups
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupUpdate'
      responses:
        '200':
          description: Group updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete group
      operationId: delete_group
      tags:
      - groups
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /groups/{group_id}/members:
    get:
      summary: List group members
      operationId: list_group_members
      tags:
      - groups
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
      - name: status
        in: query
        schema:
          $ref: '#/components/schemas/MemberStatus'
      responses:
        '200':
          description: List of members
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedMemberList'
    post:
      summary: Add member to group
      operationId: add_group_member
      tags:
      - groups
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberAdd'
      responses:
        '201':
          description: Member added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
  /groups/{group_id}/members/{user_id}:
    patch:
      summary: Update member
      operationId: update_group_member
      tags:
      - groups
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
      - name: user_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberUpdate'
      responses:
        '200':
          description: Member updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
    delete:
      summary: Remove member from group
      operationId: remove_group_member
      tags:
      - groups
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
      - name: user_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Member removed
...
  /admin/posts:
    delete:
      summary: Batch delete posts
      operationId: batch_delete_posts
      tags:
      - admin
      security:
      - oauth2:
        - admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchIds'
      responses:
        '200':
          description: Batch result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchResult'
  /admin/posts/status:
    patch:
      summary: Batch update post status
      operationId: batch_update_post_status
      tags:
      - admin
      security:
      - oauth2:
        - admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchStatusUpdate'
      responses:
        '200':
          description: Batch result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchResult'
  /admin/users/role:
    patch:
      summary: Batch update user roles
      operationId: batch_update_user_roles
      tags:
      - admin
      security:
      - oauth2:
        - admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchRoleUpdate'
      responses:
        '200':
          description: Batch result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchResult'
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: /oauth/authorize
          tokenUrl: /oauth/token
          scopes:
            read: Read access to resources
            write: Write access to resources
            admin: Admin access for batch operations
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              type: string
              example: NOT_FOUND
            message:
              type: string
              example: Resource not found
            details:
              type: object
              additionalProperties: true
...
    PostStatus:
      type: string
      enum:
      - draft
      - pending_review
      - published
      - rejected
      description: Post status
    UserRole:
      type: string
      enum:
      - participant
      - organizer
      - admin
      description: User role
    GroupVisibility:
      type: string
      enum:
      - public
      - private
      description: Group visibility
...
    MemberRole:
      type: string
      enum:
      - owner
      - admin
      - member
      description: Group member role
    MemberStatus:
      type: string
      enum:
      - pending
      - accepted
      - rejected
      description: Group membership status
...
    User:
      type: object
      required:
      - id
      - username
      - email
      - role
      - created_at
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        display_name:
          type: string
        avatar_url:
          type: string
          format: uri
        bio:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
    GroupCreate:
      type: object
      required:
      - name
      properties:
        name:
          type: string
        description:
          type: string
        visibility:
          $ref: '#/components/schemas/GroupVisibility'
        max_members:
          type: integer
        require_approval:
          type: boolean
          default: false
    GroupUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        visibility:
          $ref: '#/components/schemas/GroupVisibility'
        max_members:
          type: integer
        require_approval:
          type: boolean
          default: false
    Group:
      type: object
      required:
      - id
      - name
      - visibility
      - created_at
      properties:
        name:
          type: string
        description:
          type: string
        visibility:
          $ref: '#/components/schemas/GroupVisibility'
        max_members:
          type: integer
        require_approval:
          type: boolean
          default: false
...
    MemberAdd:
      type: object
      required:
      - user_id
      properties:
        user_id:
          type: string
        role:
          $ref: '#/components/schemas/MemberRole'
    MemberUpdate:
      type: object
      properties:
        role:
          $ref: '#/components/schemas/MemberRole'
        status:
          $ref: '#/components/schemas/MemberStatus'
    Member:
      type: object
      required:
      - user_id
      - role
      - status
      properties:
        user_id:
          type: string
        user:
          $ref: '#/components/schemas/User'
        role:
          $ref: '#/components/schemas/MemberRole'
        status:
          $ref: '#/components/schemas/MemberStatus'
        joined_at:
          type: string
          format: date-time
        status_changed_at:
          type: string
          format: date-time
...
    PaginatedGroupList:
      type: object
      required:
      - items
      - total
      - skip
      - limit
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Group'
        total:
          type: integer
          description: Total number of records
        skip:
          type: integer
        limit:
          type: integer
...
    PaginatedMemberList:
      type: object
      required:
      - items
      - total
      - skip
      - limit
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Member'
        total:
          type: integer
          description: Total number of records
        skip:
          type: integer
        limit:
          type: integer
    BatchIds:
      type: object
      required:
      - ids
      properties:
        ids:
          type: array
          items:
            type: string
          minItems: 1
          maxItems: 100
    BatchStatusUpdate:
      type: object
      required:
      - ids
      - status
      properties:
        ids:
          type: array
          items:
            type: string
          minItems: 1
          maxItems: 100
        status:
          $ref: '#/components/schemas/PostStatus'
    BatchRoleUpdate:
      type: object
      required:
      - ids
      - role
      properties:
        ids:
          type: array
          items:
            type: string
          minItems: 1
          maxItems: 100
        role:
          $ref: '#/components/schemas/UserRole'
    BatchResult:
      type: object
      required:
      - success_count
      - failed_count
      properties:
        success_count:
          type: integer
        failed_count:
          type: integer
        failed_ids:
          type: array
          items:
            type: string
        errors:
          type: object
          additionalProperties:
            type: string
    LoginRequest:
      type: object
      required:
      - email
      - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
    LoginResponse:
      type: object
      required:
      - access_token
      - token_type
      - user
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          default: bearer
        expires_in:
          type: integer
          description: Token expiration time in seconds
        user:
          $ref: '#/components/schemas/User'
    RefreshRequest:
      type: object
      required:
      - refresh_token
      properties:
        refresh_token:
          type: string
...
  parameters:
    SkipParam:
      name: skip
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
      description: Number of records to skip
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      description: Maximum number of records to return
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
security:
- oauth2:
  - read
  - write